% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyecdm-package.R
\docType{package}
\name{shinyecdm-package}
\alias{shinyecdm}
\alias{shinyecdm-package}
\title{shinyecdm: Shiny Graphical User Interface for the ecdm Package}
\description{
Provides a web interface for interacting with the 'ecdm' R package
using the Shiny framework.
}
\details{
The \code{shinyecdm} package provides a way to encase the estimation routines
in a web browser.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/shinyecdm}
  \item Report bugs at \url{https://github.com/tmsalab/shinyecdm/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
